+++ .github/workflows/build-and-test.yml
@@ -1,0 +1,30 @@
+name: build and test senaite.ast
+on:
+  - push
+  - pull_request
+env:
+  PLONE_VERSION: "5.2"
+jobs:
+  build-and-test:
+    runs-on: 'ubuntu-20.04'
+    container:
+      image: python:2.7.18-buster
+    steps:
+      - uses: actions/checkout@v3
+      - name: cache eggs
+        uses: actions/cache@v3
+        with:
+          key: eggs-cache-${{ hashFiles('buildout.cfg', 'requirements.txt') }}
+          path: |
+            eggs/
+      - name: install
+        run: |
+          pip install virtualenv
+          virtualenv -p `which python` .
+          bin/pip install --upgrade pip
+          bin/pip install -r requirements.txt
+          bin/buildout -N -t 3 annotate
+          bin/buildout -N -t 3
+      - name: test
+        run: |
+          bin/test -s senaite.ast.tests
