--- .travis.yml
@@ -1,21 +1,0 @@
-version: ~> 1.0
-language: python
-os: linux
-dist: xenial
-cache:
-  pip: true
-  directories:
-    - eggs
-matrix:
-  fast_finish: true
-  include:
-    - python: "2.7"
-      env: PLONE_VERSION="5.2"
-before_install:
-  - virtualenv -p `which python` .
-  - bin/pip install -r requirements.txt
-  - bin/buildout -N -t 3 annotate
-install:
-  - bin/buildout -N -t 3
-script:
-  - bin/test -s senaite.ast
