+++ buildout.cfg
@@ -1,105 +1,18 @@
 [buildout]
-index = https://pypi.org/simple/
-extends = https://dist.plone.org/release/5.2-latest/versions.cfg
-find-links =
-    https://dist.plone.org/release/5.2-latest/
-    https://dist.plone.org/thirdparty/
+extends = https://raw.githubusercontent.com/senaite/senaite.core/2.x/buildout.cfg
+
+package-name = senaite.ast
 
 parts =
     instance
     test
     omelette
-    i18ndude
-    zopepy
     update_translations
     write_code_headers
 
-eggs =
-    senaite.core
-    senaite.app.listing
-    senaite.app.spotlight
-    senaite.app.supermodel
-    senaite.impress
-    senaite.jsonapi
-    senaite.lims
-    senaite.abx
-    senaite.microorganism
-    senaite.ast
-    plone.reload
-    Products.PrintingMailHost
-
-extensions = mr.developer
-
-package-name = senaite.ast
-
-versions = versions
-show-picked-versions = true
-
-plone-user = admin:admin
-
-develop = .
-sources = sources
-auto-checkout = *
-
 [sources]
-senaite.core = git https://github.com/senaite/senaite.core.git branch=2.x
-senaite.app.listing = git https://github.com/senaite/senaite.app.listing.git branch=2.x
-senaite.app.spotlight = git https://github.com/senaite/senaite.app.spotlight.git branch=2.x
-senaite.app.supermodel = git https://github.com/senaite/senaite.app.supermodel.git branch=2.x
-senaite.impress = git https://github.com/senaite/senaite.impress.git branch=2.x
-senaite.jsonapi = git https://github.com/senaite/senaite.jsonapi.git branch=2.x
-senaite.lims = git https://github.com/senaite/senaite.lims.git branch=2.x
-senaite.abx = git https://github.com/senaite/senaite.abx.git
-senaite.microorganism = git https://github.com/senaite/senaite.microorganism.git
-
-[instance]
-recipe = plone.recipe.zope2instance
-http-address = 127.0.0.1:8080
-user = ${buildout:plone-user}
-wsgi = on
-eggs =
-    Plone
-    plone.app.upgrade
-    ${buildout:package-name}
-    ${buildout:eggs}
-deprecation-warnings = on
-environment-vars =
-    zope_i18n_compile_mo_files true
-zcml =
-
-[i18ndude]
-unzip = true
-recipe = zc.recipe.egg
-eggs = i18ndude
-
-[update_translations]
-recipe = collective.recipe.template
-output = ${buildout:directory}/bin/update_translations
-input = ${buildout:directory}/templates/update_translations.in
-mode = 755
-
-[write_code_headers]
-recipe = collective.recipe.template
-output = ${buildout:directory}/bin/write_code_headers
-input = ${buildout:directory}/templates/write_code_headers.py.in
-mode = 755
+senaite.core = git https://github.com/senaite/senaite.core.git
 
 [test]
-recipe = zc.recipe.testrunner
-defaults = ['--auto-color', '--auto-progress']
 eggs =
     senaite.ast [test]
-
-[omelette]
-recipe = collective.recipe.omelette
-eggs = ${buildout:eggs}
-
-[zopepy]
-recipe = zc.recipe.egg
-eggs = ${instance:eggs}
-interpreter = zopepy
-scripts = zopepy
-
-[versions]
-setuptools =
-zc.buildout =
