+++ requirements.txt
@@ -1,3 +1,3 @@
-setuptools==42.0.2
+setuptools==44.1.1
 zc.buildout==2.13.3
 wheel
