--- src/senaite/ast/tests/layers.py
@@ -1,75 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# This file is part of SENAITE.AST.
-#
-# SENAITE.AST is free software: you can redistribute it and/or modify it under
-# the terms of the GNU General Public License as published by the Free
-# Software Foundation, version 2.
-#
-# This program is distributed in the hope that it will be useful, but WITHOUT
-# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
-# FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
-# details.
-#
-# You should have received a copy of the GNU General Public License along with
-# this program; if not, write to the Free Software Foundation, Inc., 51
-# Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
-#
-# Copyright 2020-2024 by it's authors.
-# Some rights reserved, see README and LICENSE.
-
-import transaction
-from plone.app.testing import applyProfile
-from plone.app.testing import FunctionalTesting
-from plone.app.testing import PLONE_FIXTURE
-from plone.app.testing import PloneSandboxLayer
-from plone.testing import zope
-
-
-class BaseLayer(PloneSandboxLayer):
-    defaultBases = (PLONE_FIXTURE,)
-
-    def setUpZope(self, app, configurationContext):
-        super(BaseLayer, self).setUpZope(app, configurationContext)
-
-        # Load ZCML
-        import bika.lims
-        import senaite.app.listing
-        import senaite.app.spotlight
-        import senaite.core
-        import senaite.abx
-        import senaite.ast
-        import senaite.impress
-        import senaite.microorganism
-
-        self.loadZCML(package=bika.lims)
-        self.loadZCML(package=senaite.core)
-        self.loadZCML(package=senaite.app.listing)
-        self.loadZCML(package=senaite.app.spotlight)
-        self.loadZCML(package=senaite.impress)
-        self.loadZCML(package=senaite.lims)
-        self.loadZCML(package=senaite.abx)
-        self.loadZCML(package=senaite.microorganism)
-        self.loadZCML(package=senaite.ast)
-
-        # Install product and call its initialize() function
-        zope.installProduct(app, "bika.lims")
-        zope.installProduct(app, "senaite.core")
-        zope.installProduct(app, "senaite.app.listing")
-        zope.installProduct(app, "senaite.app.spotlight")
-        zope.installProduct(app, "senaite.impress")
-        zope.installProduct(app, "senaite.lims")
-        zope.installProduct(app, "senaite.abx")
-        zope.installProduct(app, "senaite.microorganism")
-        zope.installProduct(app, "senaite.ast")
-
-    def setUpPloneSite(self, portal):
-        # Install into Plone site using portal_setup
-        applyProfile(portal, "senaite.core:default")
-        applyProfile(portal, "senaite.ast:default")
-        transaction.commit()
-
-
-BASE_LAYER_FIXTURE = BaseLayer()
-BASE_TESTING = FunctionalTesting(
-    bases=(BASE_LAYER_FIXTURE,), name="SENAITE.AST:BaseTesting")
