--- travis.cfg
@@ -1,22 +1,0 @@
-[buildout]
-extends =
-    buildout.cfg
-
-parts =
-    instance
-    test
-
-[test]
-recipe = collective.xmltestreport
-defaults = ['--auto-color', '--auto-progress']
-eggs =
-    senaite.ast [test]
-
-# network speedup
-socket-timeout = 5
-allow-hosts =
-    *.python.org
-    docutils.sourceforge.net
-    prdownloads.sourceforge.net
-    effbot.org
-    dist.plone.org
